refactor: fix Type Hint
