+++ zae_engine/trainer/addons/mpu.py
@@ -8,7 +8,7 @@ from torch.nn.parallel import DistributedDataParallel as DDP
 from torch.utils.data import DistributedSampler
 
 from .core import AddOnBase
-from .. import T
+from .._trainer import T
 
 
 class MultiGPUAddon(AddOnBase):
